Code improvements

+++ polocloud-master/src/main/java/de/polocloud/bootstrap/listener/ChannelInactiveListener.java
@@ -3,11 +3,11 @@ package de.polocloud.bootstrap.listener;
 import de.polocloud.api.PoloCloudAPI;
 import de.polocloud.api.event.handling.IEventHandler;
 import de.polocloud.api.event.impl.net.ChannelInactiveEvent;
-import de.polocloud.api.gameserver.IGameServer;
-import de.polocloud.api.wrapper.IWrapper;
+import de.polocloud.api.gameserver.base.IGameServer;
+import de.polocloud.api.wrapper.base.IWrapper;
 import de.polocloud.api.wrapper.IWrapperManager;
-import de.polocloud.logger.log.Logger;
-import de.polocloud.logger.log.types.LoggerType;
+import de.polocloud.api.logger.PoloLogger;
+import de.polocloud.api.logger.helper.LogLevel;
 import io.netty.channel.ChannelHandlerContext;
 
 import java.util.LinkedList;
@@ -17,6 +17,7 @@ public class ChannelInactiveListener implements IEventHandler<ChannelInactiveEve
 
     @Override
     public void handleEvent(ChannelInactiveEvent event) {
+
         ChannelHandlerContext ctx = event.getChx();
 
         IWrapperManager wrapperManager = PoloCloudAPI.getInstance().getWrapperManager();
@@ -26,16 +27,19 @@ public class ChannelInactiveListener implements IEventHandler<ChannelInactiveEve
                 continue;
             }
             if (wrapper.ctx().channel().id().asLongText().equalsIgnoreCase(ctx.channel().id().asLongText())) {
-                Logger.log(LoggerType.INFO, "§7The Wrapper §7'§3" + wrapper.getName() + "§7' §cdisconnected§7!");
+                PoloLogger.print(LogLevel.INFO, "§7The Wrapper §7'§3" + wrapper.getName() + "§7' §cdisconnected§7!");
 
                 //Stopping all servers from the wrapper
-                for (IGameServer server : wrapper.getServers()) {
-                    server.terminate();
+                if (!wrapper.getServers().isEmpty()) {
+                    for (IGameServer server : wrapper.getServers()) {
+                        if (server.isRegistered()) {
+                            server.terminate();
+                        }
+                    }
                 }
 
                 //Finally unregistering
                 wrapperManager.unregisterWrapper(wrapper);
-                wrapperManager.syncCache();
             }
         }
